fix c warnings

+++ vlib/builtin/map_test.v
@@ -753,7 +753,7 @@ fn test_in_map_literal() {
 
 fn test_byte_keys() {
 	mut m := map[byte]byte{}
-	byte_max := 255
+	byte_max := byte(255)
 	for i in byte(0) .. byte_max {
 		m[i] = i
 		assert m[i] == i
@@ -779,7 +779,7 @@ fn test_byte_keys() {
 
 fn test_i16_keys() {
 	mut m := map[i16]i16{}
-	end := 1000
+	end := i16(1000)
 	for i in i16(0) .. end {
 		m[i] = i
 		assert m[i] == i
@@ -805,7 +805,7 @@ fn test_i16_keys() {
 
 fn test_u16_keys() {
 	mut m := map[u16]u16{}
-	end := 1000
+	end := u16(1000)
 	for i in u16(0) .. end {
 		m[i] = i
 		assert m[i] == i
@@ -831,7 +831,7 @@ fn test_u16_keys() {
 
 fn test_u32_keys() {
 	mut m := map[u32]u32{}
-	end := 1000
+	end := u32(1000)
 	for i in u32(0) .. end {
 		m[i] = i
 		assert m[i] == i
@@ -883,7 +883,7 @@ fn test_int_keys2() {
 
 fn test_i64_keys() {
 	mut m := map[i64]i64{}
-	end := 1000
+	end := i64(1000)
 	for i in i64(0) .. end {
 		m[i] = i
 		assert m[i] == i
@@ -909,7 +909,7 @@ fn test_i64_keys() {
 
 fn test_u64_keys() {
 	mut m := map[u64]u64{}
-	end := 1000
+	end := u64(1000)
 	for i in u64(0) .. end {
 		m[i] = i
 		assert m[i] == i
